pyproject.toml
[tool.poetry]
name = "webstack"
version = "0.1.0"
description = ""
authors = ["Sam Haaf <samhaaf@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.8"
boto3 = "*"
hjson = "^3.0.2"
argparse = "^1.4.0"
alembic = "^1.6.5"
psycopg2-binary = "^2.9.1"
sqlalchemy-aurora-data-api = "^0.2.7"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
